completed config migration to use storage
