+++ path_config.go
@@ -2,6 +2,7 @@ package keyfactor
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/hashicorp/vault/sdk/framework"
@@ -9,7 +10,7 @@ import (
 )
 
 const (
-	configStoragePath = "config"
+	configPath = "config"
 )
 
 // keyfactorConfig includes the minimum configuration
@@ -22,66 +23,82 @@ type keyfactorConfig struct {
 	CertAuthority string `json:"ca"`
 }
 
+func (b *keyfactorBackend) config(ctx context.Context, s logical.Storage) (*keyfactorConfig, error) {
+	b.lock.Lock()
+	defer b.lock.Unlock()
+
+	if b.cachedConfig != nil {
+		return b.cachedConfig, nil
+	}
+
+	entry, err := s.Get(ctx, configPath)
+	if err != nil {
+		return nil, err
+	}
+	if entry == nil {
+		return nil, nil
+	}
+
+	config := &keyfactorConfig{}
+	if err := entry.DecodeJSON(config); err != nil {
+		return nil, err
+	}
+
+	b.cachedConfig = config
+
+	return config, nil
+}
+
 // pathConfig extends the Vault API with a `/config`
 // endpoint for the backend. You can choose whether
 // or not certain attributes should be displayed,
 // required, and named. For example, password
 // is marked as sensitive and will not be output
 // when you read the configuration.
-func pathConfig(b *keyfactorBackend) *framework.Path {
-	return &framework.Path{
-		Pattern: "config",
-		Fields: map[string]*framework.FieldSchema{
-			"username": {
-				Type:        framework.TypeString,
-				Description: "The Keyfactor user name for authenticating with the platform.",
-				Required:    true,
-				DisplayAttrs: &framework.DisplayAttributes{
-					Name:      "Username",
-					Sensitive: false,
+func pathConfig(b *keyfactorBackend) []*framework.Path {
+	return []*framework.Path{
+		{
+			Pattern: `config`,
+			Fields: map[string]*framework.FieldSchema{
+				"username": {
+					Type:        framework.TypeString,
+					Description: "The Keyfactor user name for authenticating with the platform.",
+					Required:    true,
+				},
+				"password": {
+					Type:        framework.TypeString,
+					Description: "The password for the Keyfactor account used for authenticating.",
+					Required:    true,
+					DisplayAttrs: &framework.DisplayAttributes{
+						Sensitive: true,
+					},
+				},
+				"url": {
+					Type:        framework.TypeString,
+					Description: "The URL for the Keyfactor platform.",
+					Required:    true,
+				},
+				"template": {
+					Type:        framework.TypeString,
+					Description: "The certificate template to use with this instance of the plugin.",
+					Required:    true,
+				},
+				"ca": {
+					Type:        framework.TypeString,
+					Description: "The certificate authority to use with this instance of the plugin",
+					Required:    true,
 				},
 			},
-			"password": {
-				Type:        framework.TypeString,
-				Description: "The password for the Keyfactor account used for authenticating.",
-				Required:    true,
-				DisplayAttrs: &framework.DisplayAttributes{
-					Name:      "Password",
-					Sensitive: true,
-				},
+
+			Callbacks: map[logical.Operation]framework.OperationFunc{
+				logical.ReadOperation:   b.pathConfigRead,
+				logical.UpdateOperation: b.pathConfigWrite,
 			},
-			"url": {
-				Type:        framework.TypeString,
-				Description: "The URL for the Keyfactor platform.",
-				Required:    true,
-				DisplayAttrs: &framework.DisplayAttributes{
-					Name:      "Url",
-					Sensitive: false,
-				},
-			},
-			"template": {
-				Type:        framework.TypeString,
-				Description: "The certificate template to use with this instance of the plugin.",
-				Required:    true,
-				DisplayAttrs: &framework.DisplayAttributes{
-					Name:      "Username",
-					Sensitive: false,
-				},
-			},
-			"ca": {
-				Type:        framework.TypeString,
-				Description: "The certificate authority to use with this instance of the plugin",
-				Required:    true,
-				DisplayAttrs: &framework.DisplayAttributes{
-					Name:      "CA",
-					Sensitive: false,
-				},
-			},
+
+			//ExistenceCheck:  b.pathConfigExistenceCheck,
+			HelpSynopsis:    pathConfigHelpSynopsis,
+			HelpDescription: pathConfigHelpDescription,
 		},
-		Operations:      map[logical.Operation]framework.OperationHandler{},
-		ExistenceCheck:  b.pathConfigExistenceCheck,
-		HelpSynopsis:    pathConfigHelpSynopsis,
-		HelpDescription: pathConfigHelpDescription,
 	}
 }
 
@@ -95,8 +112,117 @@ func (b *keyfactorBackend) pathConfigExistenceCheck(ctx context.Context, req *lo
 	return out != nil, nil
 }
 
+// pathConfigRead reads the configuration and outputs non-sensitive information.
+func (b *keyfactorBackend) pathConfigRead(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+	config, err := b.config(ctx, req.Storage)
+	if err != nil {
+		return nil, err
+	}
+	if config == nil {
+		return nil, nil
+	}
+
+	return &logical.Response{
+		Data: map[string]interface{}{
+			"url":      config.KeyfactorUrl,
+			"username": config.Username,
+			"password": config.Password,
+			"CA":       config.CertAuthority,
+			"template": config.CertTemplate,
+		},
+	}, nil
+}
+
+// pathConfigWrite updates the configuration for the backend
+func (b *keyfactorBackend) pathConfigWrite(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+	b.Logger().Debug("Calling pathConfigWrite")
+
+	newConfig := &keyfactorConfig{
+		KeyfactorUrl:  data.Get("url").(string),
+		Username:      data.Get("username").(string),
+		Password:      data.Get("password").(string),
+		CertAuthority: data.Get("ca").(string),
+		CertTemplate:  data.Get("template").(string),
+	}
+
+	// Check if the config already exists, to determine if this is a create or
+	// an update, since req.Operation is always 'update' in this handler, and
+	// there's no existence check defined.
+	existingConfig, err := b.config(ctx, req.Storage)
+	if err != nil {
+		return nil, err
+	}
+
+	if existingConfig == nil {
+		existingConfig = newConfig
+	}
+
+	createOperation := (req.Operation == logical.CreateOperation)
+
+	if newConfig == nil {
+		if !createOperation {
+			return nil, errors.New("config not found during update operation")
+		}
+	}
+
+	if username, ok := data.GetOk("username"); ok {
+		existingConfig.Username = username.(string)
+	} else if !ok && createOperation {
+		return nil, fmt.Errorf("missing username in configuration")
+	}
+
+	if url, ok := data.GetOk("url"); ok {
+		existingConfig.KeyfactorUrl = url.(string)
+	} else if !ok && createOperation {
+		return nil, fmt.Errorf("missing url in configuration")
+	}
+
+	if password, ok := data.GetOk("password"); ok {
+		existingConfig.Password = password.(string)
+	} else if !ok && createOperation {
+		return nil, fmt.Errorf("missing password in configuration")
+	}
+
+	if ca, ok := data.GetOk("ca"); ok {
+		existingConfig.CertAuthority = ca.(string)
+	} else if !ok && createOperation {
+		return nil, fmt.Errorf("missing Certificate Authority in configuration")
+	}
+
+	if template, ok := data.GetOk("template"); ok {
+		existingConfig.CertTemplate = template.(string)
+	} else if !ok && createOperation {
+		return nil, fmt.Errorf("missing Certificate Template in configuration")
+	}
+
+	entry, err := logical.StorageEntryJSON(configPath, existingConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := req.Storage.Put(ctx, entry); err != nil {
+		return nil, err
+	}
+
+	// reset the client so the next invocation will pick up the new configuration
+	b.reset()
+
+	return nil, nil
+}
+
+// pathConfigDelete removes the configuration for the backend
+func (b *keyfactorBackend) pathConfigDelete(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+	err := req.Storage.Delete(ctx, configPath)
+
+	if err == nil {
+		b.reset()
+	}
+
+	return nil, err
+}
+
 func getConfig(ctx context.Context, s logical.Storage) (*keyfactorConfig, error) {
-	entry, err := s.Get(ctx, configStoragePath)
+	entry, err := s.Get(ctx, configPath)
 	if err != nil {
 		return nil, err
 	}
@@ -120,4 +246,12 @@ const pathConfigHelpSynopsis = `Configure the Keyfactor Secrets Engine backend.`
 // pathConfigHelpDescription describes the help text for the configuration
 const pathConfigHelpDescription = `
 The Keyfactor secret backend requires credentials in order to connect to the Keyfactor platform.
+`
+
+const pathConfigDesc = `
+The Keyfactor Secrets Engine plugin requires the following values to be defined:
+  username - Keyfactor user that the plugin will use for authenticating
+	password - Keyfactor user password
+	url - url of the Keyfactor platform with no trailing slashes (ie: https://keyfactor.lab)
+	ca - the certificate authority in the format <hostname\\\\logical name>
 `
