+++ path_fetch.go
@@ -1,48 +1,26 @@
 package keyfactor
 
-import (
-	"context"
-	"encoding/pem"
-	"fmt"
-	"strings"
+// import (
+// 	"context"
+// 	"encoding/pem"
+// 	"fmt"
+// 	"strings"
 
-	"github.com/hashicorp/vault/sdk/framework"
-	"github.com/hashicorp/vault/sdk/helper/errutil"
-	"github.com/hashicorp/vault/sdk/logical"
-)
+// 	"github.com/hashicorp/vault/sdk/framework"
+// 	"github.com/hashicorp/vault/sdk/helper/errutil"
+// 	"github.com/hashicorp/vault/sdk/logical"
+// )
 
-// Returns the CA in raw format
-func pathFetchCA(b *keyfactorBackend) *framework.Path {
-	return &framework.Path{
-		Pattern: `ca(/pem)?`,
+// // func pathCA(b *keyfactorBackend) *framework.Path {
+// // 	return []*framework.Path{
+// // 		pathFetchCA(&b),
+// // 	}
+// // }
 
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.ReadOperation: b.pathFetchRead,
-		},
-
-		HelpSynopsis:    pathFetchHelpSyn,
-		HelpDescription: pathFetchHelpDesc,
-	}
-}
-
-// Returns the CA chain
-func pathFetchCAChain(b *keyfactorBackend) *framework.Path {
-	return &framework.Path{
-		Pattern: `ca_chain(/pem)?`,
-
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.ReadOperation: b.pathFetchRead,
-		},
-
-		HelpSynopsis:    pathFetchHelpSyn,
-		HelpDescription: pathFetchHelpDesc,
-	}
-}
-
-// // Returns the CRL in raw format
-// func pathFetchCRL(b *backend) *framework.Path {
+// // Returns the CA in raw format
+// func pathFetchCA(b *keyfactorBackend) *framework.Path {
 // 	return &framework.Path{
-// 		Pattern: `crl(/pem)?`,
+// 		Pattern: `ca(/pem)?`,
 
 // 		Callbacks: map[logical.Operation]framework.OperationFunc{
 // 			logical.ReadOperation: b.pathFetchRead,
@@ -53,32 +31,10 @@ func pathFetchCAChain(b *keyfactorBackend) *framework.Path {
 // 	}
 // }
 
-// Returns any valid (non-revoked) cert. Since "ca" fits the pattern, this path
-// also handles returning the CA cert in a non-raw format.
-func pathFetchValid(b *keyfactorBackend) *framework.Path {
-	return &framework.Path{
-		Pattern: `cert/(?P<serial>[0-9A-Fa-f-:]+)`,
-		Fields: map[string]*framework.FieldSchema{
-			"serial": {
-				Type: framework.TypeString,
-				Description: `Certificate serial number, in colon- or
-hyphen-separated octal`,
-			},
-		},
-
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.ReadOperation: b.pathFetchRead,
-		},
-
-		HelpSynopsis:    pathFetchHelpSyn,
-		HelpDescription: pathFetchHelpDesc,
-	}
-}
-
-// This returns the CRL in a non-raw format
-// func pathFetchCRLViaCertPath(b *backend) *framework.Path {
+// // Returns the CA chain
+// func pathFetchCAChain(b *keyfactorBackend) *framework.Path {
 // 	return &framework.Path{
-// 		Pattern: `cert/crl`,
+// 		Pattern: `ca_chain(/pem)?`,
 
 // 		Callbacks: map[logical.Operation]framework.OperationFunc{
 // 			logical.ReadOperation: b.pathFetchRead,
@@ -89,196 +45,246 @@ hyphen-separated octal`,
 // 	}
 // }
 
-// This returns the list of serial numbers for certs
-func pathFetchListCerts(b *keyfactorBackend) *framework.Path {
-	return &framework.Path{
-		Pattern: "certs/?$",
+// // // Returns the CRL in raw format
+// // func pathFetchCRL(b *backend) *framework.Path {
+// // 	return &framework.Path{
+// // 		Pattern: `crl(/pem)?`,
 
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.ListOperation: b.pathFetchCertList,
-		},
+// // 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// // 			logical.ReadOperation: b.pathFetchRead,
+// // 		},
 
-		HelpSynopsis:    pathFetchHelpSyn,
-		HelpDescription: pathFetchHelpDesc,
-	}
-}
+// // 		HelpSynopsis:    pathFetchHelpSyn,
+// // 		HelpDescription: pathFetchHelpDesc,
+// // 	}
+// // }
 
-func (b *keyfactorBackend) pathFetchCertList(ctx context.Context, req *logical.Request, data *framework.FieldData) (response *logical.Response, retErr error) {
-	entries, err := req.Storage.List(ctx, "certs/")
-	if err != nil {
-		return nil, err
-	}
+// // Returns any valid (non-revoked) cert. Since "ca" fits the pattern, this path
+// // also handles returning the CA cert in a non-raw format.
+// func pathFetchValid(b *keyfactorBackend) *framework.Path {
+// 	return &framework.Path{
+// 		Pattern: `cert/(?P<serial>[0-9A-Fa-f-:]+)`,
+// 		Fields: map[string]*framework.FieldSchema{
+// 			"serial": {
+// 				Type: framework.TypeString,
+// 				Description: `Certificate serial number, in colon- or
+// hyphen-separated octal`,
+// 			},
+// 		},
 
-	return logical.ListResponse(entries), nil
-}
+// 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// 			logical.ReadOperation: b.pathFetchRead,
+// 		},
 
-func (b *keyfactorBackend) pathFetchRead(ctx context.Context, req *logical.Request, data *framework.FieldData) (response *logical.Response, retErr error) {
-	var serial, pemType, contentType string
-	var certEntry, revokedEntry *logical.StorageEntry
-	var funcErr error
-	var certificate []byte
-	var revocationTime int64
-	response = &logical.Response{
-		Data: map[string]interface{}{},
-	}
+// 		HelpSynopsis:    pathFetchHelpSyn,
+// 		HelpDescription: pathFetchHelpDesc,
+// 	}
+// }
 
-	// Some of these need to return raw and some non-raw;
-	// this is basically handled by setting contentType or not.
-	// Errors don't cause an immediate exit, because the raw
-	// paths still need to return raw output.
-	b.Logger().Debug("fetching cert, path = " + req.Path)
+// // This returns the CRL in a non-raw format
+// // func pathFetchCRLViaCertPath(b *backend) *framework.Path {
+// // 	return &framework.Path{
+// // 		Pattern: `cert/crl`,
 
-	switch {
-	case req.Path == "ca" || req.Path == "ca/pem":
-		serial = "ca"
-		contentType = "application/pkix-cert"
-		if req.Path == "ca/pem" {
-			pemType = "CERTIFICATE"
-		}
-	case req.Path == "ca_chain" || req.Path == "cert/ca_chain":
-		serial = "ca_chain"
-		if req.Path == "cert/ca_chain" {
-			contentType = "application/pkix-cert"
-		}
-	case req.Path == "crl" || req.Path == "crl/pem":
-		serial = "crl"
-		contentType = "application/pkix-crl"
-		if req.Path == "crl/pem" {
-			pemType = "X509 CRL"
-		}
-	case req.Path == "cert/crl":
-		serial = "crl"
-		pemType = "X509 CRL"
-	default:
-		serial = data.Get("serial").(string)
-		pemType = "CERTIFICATE"
-	}
-	if len(serial) == 0 {
-		response = logical.ErrorResponse("The serial number must be provided")
-		goto reply
-	}
+// // 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// // 			logical.ReadOperation: b.pathFetchRead,
+// // 		},
 
-	// if serial == "ca_chain" || serial == "ca" {
-	// 	caInfo, err := fetchCAInfo(ctx, req, b)
-	// 	switch err.(type) {
-	// 	case errutil.UserError:
-	// 		response = logical.ErrorResponse(err.Error())
-	// 		goto reply
-	// 	case errutil.InternalError:
-	// 		retErr = err
-	// 		goto reply
-	// 	}
+// // 		HelpSynopsis:    pathFetchHelpSyn,
+// // 		HelpDescription: pathFetchHelpDesc,
+// // 	}
+// // }
 
-	// 	caChain := caInfo.GetCAChain()
-	// 	var certStr string
-	// 	for _, ca := range caChain {
-	// 		block := pem.Block{
-	// 			Type:  "CERTIFICATE",
-	// 			Bytes: ca.Bytes,
-	// 		}
-	// 		certStr = strings.Join([]string{certStr, strings.TrimSpace(string(pem.EncodeToMemory(&block)))}, "\n")
-	// 	}
-	// 	certificate = []byte(strings.TrimSpace(certStr))
-	// 	goto reply
-	// }
-	b.Logger().Debug("fetching certificate; serial = " + serial)
+// // This returns the list of serial numbers for certs
+// func pathFetchListCerts(b *keyfactorBackend) *framework.Path {
+// 	return &framework.Path{
+// 		Pattern: "certs/?$",
 
-	if serial == "ca" {
-		return fetchCAInfo(ctx, req, b)
-	}
+// 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// 			logical.ListOperation: b.pathFetchCertList,
+// 		},
 
-	if serial == "ca_chain" {
-		return fetchCaChainInfo(ctx, req, b)
-	}
+// 		HelpSynopsis:    pathFetchHelpSyn,
+// 		HelpDescription: pathFetchHelpDesc,
+// 	}
+// }
 
-	certEntry, funcErr = fetchCertBySerial(ctx, req, req.Path, serial)
-	if funcErr != nil {
-		switch funcErr.(type) {
-		case errutil.UserError:
-			response = logical.ErrorResponse(funcErr.Error())
-			goto reply
-		case errutil.InternalError:
-			retErr = funcErr
-			goto reply
-		}
-	}
-	if certEntry == nil {
-		response = nil
-		goto reply
-	}
+// // func (b *keyfactorBackend) pathFetchCertList(ctx context.Context, req *logical.Request, data *framework.FieldData) (response *logical.Response, retErr error) {
+// // 	entries, err := req.Storage.List(ctx, "certs/")
+// // 	if err != nil {
+// // 		return nil, err
+// // 	}
 
-	certificate = certEntry.Value
+// // 	return logical.ListResponse(entries), nil
+// // }
 
-	if len(pemType) != 0 {
-		block := pem.Block{
-			Type:  pemType,
-			Bytes: certEntry.Value,
-		}
-		// This is convoluted on purpose to ensure that we don't have trailing
-		// newlines via various paths
-		certificate = []byte(strings.TrimSpace(string(pem.EncodeToMemory(&block))))
-	}
+// func (b *keyfactorBackend) pathFetchRead(ctx context.Context, req *logical.Request, data *framework.FieldData) (response *logical.Response, retErr error) {
+// 	var serial, pemType, contentType string
+// 	var certEntry, revokedEntry *logical.StorageEntry
+// 	var funcErr error
+// 	var certificate []byte
+// 	var revocationTime int64
+// 	response = &logical.Response{
+// 		Data: map[string]interface{}{},
+// 	}
 
-	revokedEntry, funcErr = fetchCertBySerial(ctx, req, "revoked/", serial)
-	if funcErr != nil {
-		switch funcErr.(type) {
-		case errutil.UserError:
-			response = logical.ErrorResponse(funcErr.Error())
-			goto reply
-		case errutil.InternalError:
-			retErr = funcErr
-			goto reply
-		}
-	}
-	if revokedEntry != nil {
-		var revInfo revocationInfo
-		err := revokedEntry.DecodeJSON(&revInfo)
-		if err != nil {
-			return logical.ErrorResponse(fmt.Sprintf("Error decoding revocation entry for serial %s: %s", serial, err)), nil
-		}
-		revocationTime = revInfo.RevocationTime
-	}
+// 	// Some of these need to return raw and some non-raw;
+// 	// this is basically handled by setting contentType or not.
+// 	// Errors don't cause an immediate exit, because the raw
+// 	// paths still need to return raw output.
+// 	b.Logger().Debug("fetching cert, path = " + req.Path)
 
-reply:
-	switch {
-	case len(contentType) != 0:
-		response = &logical.Response{
-			Data: map[string]interface{}{
-				logical.HTTPContentType: contentType,
-				logical.HTTPRawBody:     certificate,
-			}}
-		if retErr != nil {
-			if b.Logger().IsWarn() {
-				b.Logger().Warn("possible error, but cannot return in raw response. Note that an empty CA probably means none was configured, and an empty CRL is possibly correct", "error", retErr)
-			}
-		}
-		retErr = nil
-		if len(certificate) > 0 {
-			response.Data[logical.HTTPStatusCode] = 200
-		} else {
-			response.Data[logical.HTTPStatusCode] = 204
-		}
-	case retErr != nil:
-		response = nil
-		return
-	case response == nil:
-		return
-	case response.IsError():
-		return response, nil
-	default:
-		response.Data["certificate"] = string(certificate)
-		response.Data["revocation_time"] = revocationTime
-	}
+// 	switch {
+// 	case req.Path == "ca" || req.Path == "ca/pem":
+// 		serial = "ca"
+// 		contentType = "application/pkix-cert"
+// 		if req.Path == "ca/pem" {
+// 			pemType = "CERTIFICATE"
+// 		}
+// 	case req.Path == "ca_chain" || req.Path == "cert/ca_chain":
+// 		serial = "ca_chain"
+// 		if req.Path == "cert/ca_chain" {
+// 			contentType = "application/pkix-cert"
+// 		}
+// 	case req.Path == "crl" || req.Path == "crl/pem":
+// 		serial = "crl"
+// 		contentType = "application/pkix-crl"
+// 		if req.Path == "crl/pem" {
+// 			pemType = "X509 CRL"
+// 		}
+// 	case req.Path == "cert/crl":
+// 		serial = "crl"
+// 		pemType = "X509 CRL"
+// 	default:
+// 		serial = data.Get("serial").(string)
+// 		pemType = "CERTIFICATE"
+// 	}
+// 	if len(serial) == 0 {
+// 		response = logical.ErrorResponse("The serial number must be provided")
+// 		goto reply
+// 	}
 
-	return
-}
+// 	// if serial == "ca_chain" || serial == "ca" {
+// 	// 	caInfo, err := fetchCAInfo(ctx, req, b)
+// 	// 	switch err.(type) {
+// 	// 	case errutil.UserError:
+// 	// 		response = logical.ErrorResponse(err.Error())
+// 	// 		goto reply
+// 	// 	case errutil.InternalError:
+// 	// 		retErr = err
+// 	// 		goto reply
+// 	// 	}
 
-const pathFetchHelpSyn = `
-Fetch a CA, CRL, CA Chain, or non-revoked certificate.
-`
+// 	// 	caChain := caInfo.GetCAChain()
+// 	// 	var certStr string
+// 	// 	for _, ca := range caChain {
+// 	// 		block := pem.Block{
+// 	// 			Type:  "CERTIFICATE",
+// 	// 			Bytes: ca.Bytes,
+// 	// 		}
+// 	// 		certStr = strings.Join([]string{certStr, strings.TrimSpace(string(pem.EncodeToMemory(&block)))}, "\n")
+// 	// 	}
+// 	// 	certificate = []byte(strings.TrimSpace(certStr))
+// 	// 	goto reply
+// 	// }
+// 	b.Logger().Debug("fetching certificate; serial = " + serial)
 
-const pathFetchHelpDesc = `
-This allows certificates to be fetched. If using the fetch/ prefix any non-revoked certificate can be fetched.
-Using "ca" or "crl" as the value fetches the appropriate information in DER encoding. Add "/pem" to either to get PEM encoding.
-Using "ca_chain" as the value fetches the certificate authority trust chain in PEM encoding.
-`
+// 	if serial == "ca" {
+// 		return fetchCAInfo(ctx, req, b)
+// 	}
+
+// 	if serial == "ca_chain" {
+// 		return fetchCaChainInfo(ctx, req, b)
+// 	}
+
+// 	certEntry, funcErr = fetchCertBySerial(ctx, req, req.Path, serial)
+// 	if funcErr != nil {
+// 		switch funcErr.(type) {
+// 		case errutil.UserError:
+// 			response = logical.ErrorResponse(funcErr.Error())
+// 			goto reply
+// 		case errutil.InternalError:
+// 			retErr = funcErr
+// 			goto reply
+// 		}
+// 	}
+// 	if certEntry == nil {
+// 		response = nil
+// 		goto reply
+// 	}
+
+// 	certificate = certEntry.Value
+
+// 	if len(pemType) != 0 {
+// 		block := pem.Block{
+// 			Type:  pemType,
+// 			Bytes: certEntry.Value,
+// 		}
+// 		// This is convoluted on purpose to ensure that we don't have trailing
+// 		// newlines via various paths
+// 		certificate = []byte(strings.TrimSpace(string(pem.EncodeToMemory(&block))))
+// 	}
+
+// 	revokedEntry, funcErr = fetchCertBySerial(ctx, req, "revoked/", serial)
+// 	if funcErr != nil {
+// 		switch funcErr.(type) {
+// 		case errutil.UserError:
+// 			response = logical.ErrorResponse(funcErr.Error())
+// 			goto reply
+// 		case errutil.InternalError:
+// 			retErr = funcErr
+// 			goto reply
+// 		}
+// 	}
+// 	if revokedEntry != nil {
+// 		var revInfo revocationInfo
+// 		err := revokedEntry.DecodeJSON(&revInfo)
+// 		if err != nil {
+// 			return logical.ErrorResponse(fmt.Sprintf("Error decoding revocation entry for serial %s: %s", serial, err)), nil
+// 		}
+// 		revocationTime = revInfo.RevocationTime
+// 	}
+
+// reply:
+// 	switch {
+// 	case len(contentType) != 0:
+// 		response = &logical.Response{
+// 			Data: map[string]interface{}{
+// 				logical.HTTPContentType: contentType,
+// 				logical.HTTPRawBody:     certificate,
+// 			}}
+// 		if retErr != nil {
+// 			if b.Logger().IsWarn() {
+// 				b.Logger().Warn("possible error, but cannot return in raw response. Note that an empty CA probably means none was configured, and an empty CRL is possibly correct", "error", retErr)
+// 			}
+// 		}
+// 		retErr = nil
+// 		if len(certificate) > 0 {
+// 			response.Data[logical.HTTPStatusCode] = 200
+// 		} else {
+// 			response.Data[logical.HTTPStatusCode] = 204
+// 		}
+// 	case retErr != nil:
+// 		response = nil
+// 		return
+// 	case response == nil:
+// 		return
+// 	case response.IsError():
+// 		return response, nil
+// 	default:
+// 		response.Data["certificate"] = string(certificate)
+// 		response.Data["revocation_time"] = revocationTime
+// 	}
+
+// 	return
+// }
+
+// const pathFetchHelpSyn = `
+// Fetch a CA, CRL, CA Chain, or non-revoked certificate.
+// `
+
+// const pathFetchHelpDesc = `
+// This allows certificates to be fetched. If using the fetch/ prefix any non-revoked certificate can be fetched.
+// Using "ca" or "crl" as the value fetches the appropriate information in DER encoding. Add "/pem" to either to get PEM encoding.
+// Using "ca_chain" as the value fetches the certificate authority trust chain in PEM encoding.
+// `
