+++ path_issue_sign.go
@@ -1,240 +1,228 @@
 package keyfactor
 
-import (
-	"context"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"strings"
+// func pathIssue(b *keyfactorBackend) *framework.Path {
+// 	ret := &framework.Path{
+// 		Pattern: "issue/" + framework.GenericNameRegex("role"),
 
-	"github.com/hashicorp/vault/sdk/framework"
-	"github.com/hashicorp/vault/sdk/helper/consts"
-	"github.com/hashicorp/vault/sdk/logical"
-)
+// 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// 			logical.UpdateOperation: b.pathIssue,
+// 		},
 
-func pathIssue(b *keyfactorBackend) *framework.Path {
-	ret := &framework.Path{
-		Pattern: "issue/" + framework.GenericNameRegex("role"),
+// 		HelpSynopsis:    pathIssueHelpSyn,
+// 		HelpDescription: pathIssueHelpDesc,
+// 	}
 
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.UpdateOperation: b.pathIssue,
-		},
+// 	ret.Fields = addNonCACommonFields(map[string]*framework.FieldSchema{})
+// 	return ret
+// }
 
-		HelpSynopsis:    pathIssueHelpSyn,
-		HelpDescription: pathIssueHelpDesc,
-	}
+// func pathSign(b *keyfactorBackend) *framework.Path {
+// 	ret := &framework.Path{
+// 		Pattern: "sign/" + framework.GenericNameRegex("role"),
 
-	ret.Fields = addNonCACommonFields(map[string]*framework.FieldSchema{})
-	return ret
-}
+// 		Callbacks: map[logical.Operation]framework.OperationFunc{
+// 			logical.UpdateOperation: b.pathSign,
+// 		},
 
-func pathSign(b *keyfactorBackend) *framework.Path {
-	ret := &framework.Path{
-		Pattern: "sign/" + framework.GenericNameRegex("role"),
+// 		HelpSynopsis:    pathSignHelpSyn,
+// 		HelpDescription: pathSignHelpDesc,
+// 	}
 
-		Callbacks: map[logical.Operation]framework.OperationFunc{
-			logical.UpdateOperation: b.pathSign,
-		},
+// 	ret.Fields = addNonCACommonFields(map[string]*framework.FieldSchema{})
 
-		HelpSynopsis:    pathSignHelpSyn,
-		HelpDescription: pathSignHelpDesc,
-	}
+// 	ret.Fields["csr"] = &framework.FieldSchema{
+// 		Type:        framework.TypeString,
+// 		Default:     "",
+// 		Description: `PEM-format CSR to be signed.`,
+// 		Required:    true,
+// 	}
 
-	ret.Fields = addNonCACommonFields(map[string]*framework.FieldSchema{})
-
-	ret.Fields["csr"] = &framework.FieldSchema{
-		Type:        framework.TypeString,
-		Default:     "",
-		Description: `PEM-format CSR to be signed.`,
-		Required:    true,
-	}
-
-	return ret
-}
+// 	return ret
+// }
 
 // pathIssue issues a certificate and private key from given parameters,
 // subject to role restrictions
-func (b *keyfactorBackend) pathIssue(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
-	roleName := data.Get("role").(string)
+// func (b *keyfactorBackend) pathIssue(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+// 	roleName := data.Get("role").(string)
 
-	// Get the role
-	role, err := b.getRole(ctx, req.Storage, roleName)
-	if err != nil {
-		return nil, err
-	}
-	if role == nil {
-		return logical.ErrorResponse(fmt.Sprintf("unknown role: %s", roleName)), nil
-	}
+// 	// Get the role
+// 	role, err := b.getRole(ctx, req.Storage, roleName)
+// 	if err != nil {
+// 		return nil, err
+// 	}
+// 	if role == nil {
+// 		return logical.ErrorResponse(fmt.Sprintf("unknown role: %s", roleName)), nil
+// 	}
 
-	if role.KeyType == "any" {
-		return logical.ErrorResponse("role key type \"any\" not allowed for issuing certificates, only signing"), nil
-	}
+// 	if role.KeyType == "any" {
+// 		return logical.ErrorResponse("role key type \"any\" not allowed for issuing certificates, only signing"), nil
+// 	}
 
-	return b.pathIssueSignCert(ctx, req, data, role)
-}
+// 	return b.pathIssueSignCert(ctx, req, data, role)
+// }
 
-// pathSign issues a certificate from a submitted CSR, subject to role
-// restrictions
-func (b *keyfactorBackend) pathSign(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
-	roleName := data.Get("role").(string)
-	csr := data.Get("csr").(string)
-	// Get the role
-	role, err := b.getRole(ctx, req.Storage, roleName)
-	if err != nil {
-		return nil, err
-	}
-	if role == nil {
-		return logical.ErrorResponse(fmt.Sprintf("unknown role: %s", roleName)), nil
-	}
+// // pathSign issues a certificate from a submitted CSR, subject to role
+// // restrictions
+// func (b *keyfactorBackend) pathSign(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+// 	roleName := data.Get("role").(string)
+// 	csr := data.Get("csr").(string)
+// 	// Get the role
+// 	role, err := b.getRole(ctx, req.Storage, roleName)
+// 	if err != nil {
+// 		return nil, err
+// 	}
+// 	if role == nil {
+// 		return logical.ErrorResponse(fmt.Sprintf("unknown role: %s", roleName)), nil
+// 	}
 
-	caName := data.Get("ca").(string)
-	templateName := data.Get("template").(string)
+// 	caName := data.Get("ca").(string)
+// 	templateName := data.Get("template").(string)
 
-	b.Logger().Debug("CA Name parameter = " + caName)
-	b.Logger().Debug("Template name parameter = " + templateName)
+// 	b.Logger().Debug("CA Name parameter = " + caName)
+// 	b.Logger().Debug("Template name parameter = " + templateName)
 
-	certs, serial, errr := b.submitCSR(ctx, req, csr, caName, templateName)
+// 	certs, serial, errr := b.submitCSR(ctx, req, csr, caName, templateName)
 
-	if errr != nil {
-		return nil, fmt.Errorf("could not sign csr: %s", errr)
-	}
-	response := &logical.Response{
-		Data: map[string]interface{}{
-			"certificate":   certs[0],
-			"issuing_ca":    certs[1],
-			"serial_number": serial,
-		},
-	}
+// 	if errr != nil {
+// 		return nil, fmt.Errorf("could not sign csr: %s", errr)
+// 	}
+// 	response := &logical.Response{
+// 		Data: map[string]interface{}{
+// 			"certificate":   certs[0],
+// 			"issuing_ca":    certs[1],
+// 			"serial_number": serial,
+// 		},
+// 	}
 
-	return response, nil
-}
+// 	return response, nil
+// }
 
-func (b *keyfactorBackend) pathIssueSignCert(ctx context.Context, req *logical.Request, data *framework.FieldData, role *roleEntry) (*logical.Response, error) {
-	// If storing the certificate and on a performance standby, forward this request on to the primary
-	if !role.NoStore && b.System().ReplicationState().HasState(consts.ReplicationPerformanceStandby) {
-		return nil, logical.ErrReadOnly
-	}
+// func (b *keyfactorBackend) pathIssueSignCert(ctx context.Context, req *logical.Request, data *framework.FieldData, role *roleEntry) (*logical.Response, error) {
+// 	// If storing the certificate and on a performance standby, forward this request on to the primary
+// 	if !role.NoStore && b.System().ReplicationState().HasState(consts.ReplicationPerformanceStandby) {
+// 		return nil, logical.ErrReadOnly
+// 	}
 
-	var ip_sans []string
-	var dns_sans []string
+// 	var ip_sans []string
+// 	var dns_sans []string
 
-	arg, _ := json.Marshal(req.Data)
-	b.Logger().Debug(string(arg))
+// 	arg, _ := json.Marshal(req.Data)
+// 	b.Logger().Debug(string(arg))
 
-	// get common name
-	cn, ok := data.GetOk("common_name")
+// 	// get common name
+// 	cn, ok := data.GetOk("common_name")
 
-	if !ok {
-		return nil, fmt.Errorf("common_name must be provided to issue certificate")
-	}
+// 	if !ok {
+// 		return nil, fmt.Errorf("common_name must be provided to issue certificate")
+// 	}
 
-	cn = cn.(string)
+// 	cn = cn.(string)
 
-	if cn == "" {
-		return nil, fmt.Errorf("common_name must be provided to issue certificate")
-	}
+// 	if cn == "" {
+// 		return nil, fmt.Errorf("common_name must be provided to issue certificate")
+// 	}
 
-	// get dns sans (required)
-	dns_sans_string, ok := data.GetOk("dns_sans")
+// 	// get dns sans (required)
+// 	dns_sans_string, ok := data.GetOk("dns_sans")
 
-	if !ok {
-		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
-	}
+// 	if !ok {
+// 		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
+// 	}
 
-	dns_sans_string = dns_sans_string.(string)
+// 	dns_sans_string = dns_sans_string.(string)
 
-	if dns_sans_string == "" {
-		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
-	}
+// 	if dns_sans_string == "" {
+// 		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
+// 	}
 
-	dns_sans = strings.Split(dns_sans_string.(string), ",")
+// 	dns_sans = strings.Split(dns_sans_string.(string), ",")
 
-	if len(dns_sans) == 0 {
-		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
-	}
+// 	if len(dns_sans) == 0 {
+// 		return nil, fmt.Errorf("dns_sans must be provided to issue certificate")
+// 	}
 
-	// get ip sans (optional)
-	ip_sans_string, ok := data.GetOk("ip_sans")
-	if ok {
-		ip_sans = strings.Split(ip_sans_string.(string), ",")
-	}
+// 	// get ip sans (optional)
+// 	ip_sans_string, ok := data.GetOk("ip_sans")
+// 	if ok {
+// 		ip_sans = strings.Split(ip_sans_string.(string), ",")
+// 	}
 
-	caName := data.Get("ca").(string)
+// 	caName := data.Get("ca").(string)
 
-	templateName := data.Get("template").(string)
+// 	templateName := data.Get("template").(string)
 
-	b.Logger().Debug("CA Name parameter = " + caName)
-	b.Logger().Debug("Template name parameter = " + templateName)
+// 	b.Logger().Debug("CA Name parameter = " + caName)
+// 	b.Logger().Debug("Template name parameter = " + templateName)
 
-	//check role permissions
-	var err_resp error
-	if strings.Contains(cn.(string), role.AllowedBaseDomain) && !role.AllowSubdomains {
-		err_resp = fmt.Errorf("sub-domains not allowed for role")
-	}
-	if role.AllowedBaseDomain == cn.(string) {
-		err_resp = fmt.Errorf("common name not allowed for provided role")
-	}
+// 	//check role permissions
+// 	var err_resp error
+// 	if strings.Contains(cn.(string), role.AllowedBaseDomain) && !role.AllowSubdomains {
+// 		err_resp = fmt.Errorf("sub-domains not allowed for role")
+// 	}
+// 	if role.AllowedBaseDomain == cn.(string) {
+// 		err_resp = fmt.Errorf("common name not allowed for provided role")
+// 	}
 
-	if err_resp != nil {
-		return nil, err_resp
-	}
+// 	if err_resp != nil {
+// 		return nil, err_resp
+// 	}
 
-	for u := range dns_sans {
-		if !strings.Contains(dns_sans[u], role.AllowedBaseDomain) || strings.Contains(dns_sans[u], role.AllowedBaseDomain) && !role.AllowSubdomains {
-			return nil, fmt.Errorf("Subject Alternative Name " + dns_sans[u] + " not allowed for provided role")
-		}
-	}
+// 	for u := range dns_sans {
+// 		if !strings.Contains(dns_sans[u], role.AllowedBaseDomain) || strings.Contains(dns_sans[u], role.AllowedBaseDomain) && !role.AllowSubdomains {
+// 			return nil, fmt.Errorf("Subject Alternative Name " + dns_sans[u] + " not allowed for provided role")
+// 		}
+// 	}
 
-	//generate and submit CSR
-	csr, key := b.generateCSR(cn.(string), ip_sans, dns_sans)
-	certs, serial, errr := b.submitCSR(ctx, req, csr, caName, templateName)
+// 	//generate and submit CSR
+// 	csr, key := b.generateCSR(cn.(string), ip_sans, dns_sans)
+// 	certs, serial, errr := b.submitCSR(ctx, req, csr, caName, templateName)
 
-	if errr != nil {
-		return nil, fmt.Errorf("could not enroll certificate: %s", errr)
-	}
+// 	if errr != nil {
+// 		return nil, fmt.Errorf("could not enroll certificate: %s", errr)
+// 	}
 
-	// Conform response to Vault PKI API
-	response := &logical.Response{
-		Data: map[string]interface{}{
-			"certificate":      certs[0],
-			"issuing_ca":       certs[1],
-			"private_key":      "-----BEGIN RSA PRIVATE KEY-----\n" + base64.StdEncoding.EncodeToString(key) + "\n-----END RSA PRIVATE KEY-----",
-			"private_key_type": "rsa",
-			"revocation_time":  0,
-			"serial_number":    serial,
-		},
-	}
+// 	// Conform response to Vault PKI API
+// 	response := &logical.Response{
+// 		Data: map[string]interface{}{
+// 			"certificate":      certs[0],
+// 			"issuing_ca":       certs[1],
+// 			"private_key":      "-----BEGIN RSA PRIVATE KEY-----\n" + base64.StdEncoding.EncodeToString(key) + "\n-----END RSA PRIVATE KEY-----",
+// 			"private_key_type": "rsa",
+// 			"revocation_time":  0,
+// 			"serial_number":    serial,
+// 		},
+// 	}
 
-	return response, nil
-}
+// 	return response, nil
+// }
 
-const pathIssueHelpSyn = `
-Request a certificate using a certain role with the provided details.
-example: vault write keyfactor/issue/<role> common_name=<cn> dns_sans=<dns sans>
-`
+// const pathIssueHelpSyn = `
+// Request a certificate using a certain role with the provided details.
+// example: vault write keyfactor/issue/<role> common_name=<cn> dns_sans=<dns sans>
+// `
 
-const pathIssueHelpDesc = `
-This path allows requesting a certificate to be issued according to the
-policy of the given role. The certificate will only be issued if the
-requested details are allowed by the role policy.
+// const pathIssueHelpDesc = `
+// This path allows requesting a certificate to be issued according to the
+// policy of the given role. The certificate will only be issued if the
+// requested details are allowed by the role policy.
 
-This path returns a certificate and a private key. If you want a workflow
-that does not expose a private key, generate a CSR locally and use the
-sign path instead.
-`
+// This path returns a certificate and a private key. If you want a workflow
+// that does not expose a private key, generate a CSR locally and use the
+// sign path instead.
+// `
 
-const pathSignHelpSyn = `
-Request certificates using a certain role with the provided details.
-example: vault write keyfactor/sign/<role> csr=<csr>
-`
+// const pathSignHelpSyn = `
+// Request certificates using a certain role with the provided details.
+// example: vault write keyfactor/sign/<role> csr=<csr>
+// `
 
-const pathSignHelpDesc = `
-This path allows requesting certificates to be issued according to the
-policy of the given role. The certificate will only be issued if the
-requested common name is allowed by the role policy.
+// const pathSignHelpDesc = `
+// This path allows requesting certificates to be issued according to the
+// policy of the given role. The certificate will only be issued if the
+// requested common name is allowed by the role policy.
 
-This path requires a CSR; if you want Vault to generate a private key
-for you, use the issue path instead.
+// This path requires a CSR; if you want Vault to generate a private key
+// for you, use the issue path instead.
 
-Note: the CSR must contain at least one DNS SANs entry.
-`
+// Note: the CSR must contain at least one DNS SANs entry.
+// `
